@@ -2,49 +2,45 @@
 # * Imports ----------------------------------------------------------------------------
 from __future__ import annotations
 
+import shlex
 import shutil
 import sys
+from functools import lru_cache, wraps
 
 # Should only use on python version > 3.10
 assert sys.version_info >= (3, 10)
 
-from dataclasses import replace  # noqa
+from dataclasses import dataclass
 from pathlib import Path
 from typing import (
-    Any,
+    Annotated,
     Callable,
     Iterator,
+    Literal,
     Sequence,
-    TypeVar,
-    cast,
+    TypeAlias,
+    TypedDict,
 )
 
-if sys.version_info > (3, 10):
-    from typing import TypeAlias
-else:
-    from typing_extensions import TypeAlias
-
-if sys.version_info > (3, 9):
-    from typing import Annotated
-else:
-    from typing_extensions import Annotated
-
-
 import nox  # type: ignore[unused-ignore,import]
-from noxopt import NoxOpt, Option, Session  # type: ignore[unused-ignore,import]
+from nox import Session
 
 # fmt: off
 sys.path.insert(0, ".")
+
+from tools.dataclass_parser import DataclassParser, add_option, option
 from tools.noxtools import (
+    Installer,
+    check_hash_path_for_change,
     combine_list_str,
+    infer_requirement_path,
+    is_conda_session,
     load_nox_config,
     open_webpage,
-    pkg_install_condaenv,
-    pkg_install_venv,
     prepend_flag,
     session_run_commands,
-    sort_like,
     update_target,
+    write_hashes,
 )
 
 sys.path.pop(0)
@@ -72,10 +68,12 @@ nox.options.envdir = f".nox/{PACKAGE_NAME}/envs"
 
 CONFIG = load_nox_config()
 
-# * Options ----------------------------------------------------------------------------
+# * Options ---------------------------------------------------------------------------
 
-PYTHON_ALL_VERSIONS = ["3.8", "3.9", "3.10", "3.11"]
-PYTHON_DEFAULT_VERSION = "3.10"
+LOCK = True
+
+PYTHON_ALL_VERSIONS = ["3.8", "3.9", "3.10", "3.11", "3.12"]
+PYTHON_DEFAULT_VERSION = "3.11"
 
 # conda/mamba
 if shutil.which("mamba"):
@@ -85,287 +83,308 @@ elif shutil.which("conda"):
 else:
     raise ValueError("neither conda or mamba found")
 
-SESSION_DEFAULT_KWS = {"python": PYTHON_DEFAULT_VERSION, "venv_backend": CONDA_BACKEND}
-SESSION_ALL_KWS = {"python": PYTHON_ALL_VERSIONS, "venv_backend": CONDA_BACKEND}
+
+class SessionOptionsDict(TypedDict, total=False):
+    """Dict for options to nox.session"""
+
+    python: str | list[str]
+    venv_backend: str
 
 
-# * noxopt -----------------------------------------------------------------------------
-group = NoxOpt(auto_tag=True)
+CONDA_DEFAULT_KWS: SessionOptionsDict = {
+    "python": PYTHON_DEFAULT_VERSION,
+    "venv_backend": CONDA_BACKEND,
+}
+CONDA_ALL_KWS: SessionOptionsDict = {
+    "python": PYTHON_ALL_VERSIONS,
+    "venv_backend": CONDA_BACKEND,
+}
 
-F = TypeVar("F", bound=Callable[..., Any])
-C: TypeAlias = Callable[[F], F]
-
-DEFAULT_SESSION = cast(C[F], group.session(**SESSION_DEFAULT_KWS))  # type: ignore
-ALL_SESSION = cast(C[F], group.session(**SESSION_ALL_KWS))  # type: ignore
-
-DEFAULT_SESSION_VENV = cast(C[F], group.session(python=PYTHON_DEFAULT_VERSION))  # type: ignore
-ALL_SESSION_VENV = cast(C[F], group.session(python=PYTHON_ALL_VERSIONS))  # type: ignore
-
-NOPYTHON_SESSION = cast(C[F], group.session(python=False))  # type: ignore
-INHERITED_SESSION_VENV = cast(C[F], group.session)  # type: ignore
-
-OPTS_OPT = Option(nargs="*", type=str)
-# SET_KERNEL_OPT = Option(type=bool, help="If True, try to set the kernel name")
-RUN_OPT = Option(
-    nargs="*",
-    type=str,
-    action="append",
-    help="run passed command_demo using `external=True` flag",
-)
-
-CMD_OPT = Option(nargs="*", type=str, help="cmd to be run")
-LOCK_OPT = Option(type=bool, help="If True, use conda-lock")
+DEFAULT_KWS: SessionOptionsDict = {"python": PYTHON_DEFAULT_VERSION}
+ALL_KWS: SessionOptionsDict = {"python": PYTHON_ALL_VERSIONS}
 
 
-def opts_annotated(**kwargs: Any):  # type: ignore[unused-ignore,no-untyped-def]
-    return Annotated["list[str]", replace(OPTS_OPT, **kwargs)]
+# * Session command line options -------------------------------------------------------
 
+OPT_TYPE: TypeAlias = list[str] | None
+RUN_TYPE: TypeAlias = list[list[str]] | None
 
-def cmd_annotated(**kwargs: Any):  # type: ignore[unused-ignore,no-untyped-def]
-    return Annotated["list[str]", replace(CMD_OPT, **kwargs)]
-
-
-def run_annotated(**kwargs: Any):  # type: ignore[unused-ignore,no-untyped-def]
-    return Annotated["list[list[str]]", replace(RUN_OPT, **kwargs)]
-
-
-LOCK_CLI = Annotated[bool, LOCK_OPT]
-RUN_CLI = Annotated["list[list[str]]", RUN_OPT]
-TEST_OPTS_CLI = opts_annotated(help="extra arguments/flags to pytest")
-DEV_EXTRAS_CLI = cmd_annotated(help="extras included in user dev environment")
-PYTHON_PATHS_CLI = cmd_annotated(help="python paths to append to PATHS")
-
-UPDATE_CLI = Annotated[
-    bool,
-    Option(
-        type=bool,
-        help="If True, force update of installed packages",
-        flags=("--update", "-U"),
+RUN_ANNO = Annotated[
+    RUN_TYPE,
+    option(
+        help="Run external commands in session.  Specify multiple times for multiple commands."
     ),
 ]
+OPT_ANNO = Annotated[OPT_TYPE, option(help="Options to command.")]
 
 
-UPDATE_PACKAGE_CLI = Annotated[
-    bool,
-    Option(
-        type=bool,
-        help="If True, and session uses package, reinstall package",
-        flags=("--update-package", "-P"),
-    ),
-]
+@dataclass
+class SessionParams(DataclassParser):
+    """Holds all cli options"""
 
-VERSION_CLI = Annotated[
-    str, Option(type=str, help="Version to substitute or check against")
-]
+    # common parameters
+    lock: bool = False
+    update: bool = add_option("--update", "-U", help="update dependencies/package")
+    update_package: bool = add_option(
+        "--update-package", "-P", help="update package only"
+    )
+    log_session: bool = add_option("--log-session", "-L")
+    version: str | None = None
 
-LOG_SESSION_CLI = Annotated[
-    bool,
-    Option(
-        type=bool,
-        help="If flag included, log python and package (if installed) version",
-    ),
-]
+    # dev
+    dev_run: RUN_ANNO = None
+
+    # bootstrap
+    bootstrap: list[str] | None = add_option(
+        "-B", "--bootstrap", help="run these sessions under isolated bootstrap session"
+    )
+
+    # config
+    dev_extras: OPT_TYPE = add_option(help="`extras` to include in dev environment")
+    python_paths: OPT_TYPE = add_option(help="paths to python executables")
+
+    # requirements
+    requirements_force: bool = False
+
+    # conda-lock
+    conda_lock_channel: OPT_TYPE = add_option(help="conda channels")
+    conda_lock_platform: list[
+        Literal["osx-64", "linux-64", "win-64", "osx-arm64", "all"]
+    ] | None = add_option(help="platform(s) to buiuld lock file for.")
+    conda_lock_include: OPT_TYPE = add_option(help="lock files to create")
+    conda_lock_run: RUN_ANNO = None
+    conda_lock_mamba: bool = False
+    conda_lock_force: bool = False
+
+    # pip-compile
+    pip_compile_force: bool = False
+
+    # test
+    test_no_pytest: bool = False
+    test_opts: OPT_TYPE = add_option(help="Options to pytest")
+    test_run: RUN_ANNO = None
+    no_cov: bool = False
+
+    # coverage
+    coverage: list[Literal["erase", "combine", "report", "html", "open"]] | None = None
+    coverage_run: RUN_ANNO = None
+    coverage_run_internal: RUN_TYPE = None
+
+    # testdist
+    testdist_run: RUN_ANNO = None
+
+    # docs
+    docs: list[
+        Literal[
+            "html",
+            "build",
+            "symlink",
+            "clean",
+            "livehtml",
+            "linkcheck",
+            "spelling",
+            "showlinks",
+            "release",
+            "open",
+            "serve",
+        ]
+    ] | None = add_option("--docs", "-d", help="doc commands")
+    docs_run: RUN_ANNO = None
+
+    # lint (pre-commit)
+    lint_run: RUN_ANNO = None
+
+    # typing
+    typing: list[
+        Literal[
+            "clean",
+            "mypy",
+            "pyright",
+            "pytype",
+            "all",
+            "mypy-notebook",
+            "pyright-notebook",
+            "typing-notebook",
+        ]
+    ] = add_option("--typing", "-m")
+    typing_run: RUN_ANNO = None
+    typing_run_internal: RUN_TYPE = add_option(
+        help="Run internal (in session) commands."
+    )
+
+    # build
+    build: list[Literal["build", "version"]] | None = None
+    build_run: RUN_ANNO = None
+    build_isolation: bool = False
+
+    # publish
+    publish: list[Literal["release", "test"]] | None = add_option("-p", "--publish")
+    publish_run: RUN_ANNO = None
+
+    # conda-recipe/grayskull
+    conda_recipe: list[Literal["recipe", "recipe-full"]] | None = None
+    conda_recipe_run: RUN_ANNO = None
+    conda_recipe_sdist_path: str | None = None
+
+    # conda-build
+    conda_build: list[Literal["build", "clean"]] | None = None
+    conda_build_run: RUN_ANNO = None
+
+
+@lru_cache
+def parse_posargs(*posargs: str) -> SessionParams:
+    """
+    Get Parser using `+` for session option prefix.
+
+    Note that using `+` allows for passing underlying `-` options
+    without escaping.
+    """
+    opts = SessionParams.from_posargs(posargs=posargs, prefix_char="+")
+    opts.lock = opts.lock or LOCK
+
+    return opts
+
+
+def add_opts(
+    func: Callable[[Session, SessionParams], None]
+) -> Callable[[Session], None]:
+    @wraps(func)
+    def wrapped(session: Session) -> None:
+        opts = parse_posargs(*session.posargs)
+        return func(session, opts)
+
+    return wrapped
 
 
 # * Environments------------------------------------------------------------------------
 # ** Dev (conda)
-@DEFAULT_SESSION
+@add_opts
 def dev(
     session: Session,
-    dev_run: RUN_CLI = [],  # noqa
-    lock: LOCK_CLI = False,
-    update: UPDATE_CLI = False,
-    update_package: UPDATE_PACKAGE_CLI = False,
-    log_session: bool = False,
+    opts: SessionParams,
 ) -> None:
-    """Create dev env using conda."""
-    # using conda
+    """Create development environment using either conda (dev) or virtualenv (dev-venv)."""
 
-    pkg_install_condaenv(
-        session=session,
-        name="dev",
-        lock=lock,
-        display_name=f"{PACKAGE_NAME}-dev",
-        install_package=True,
-        update=update,
-        update_package=update_package,
-        log_session=log_session,
+    (
+        Installer.from_envname(
+            session=session,
+            envname="dev",
+            lock=opts.lock,
+            update=opts.update,
+            package=".",
+        )
+        .install_all(
+            update_package=opts.update_package,
+            log_session=opts.log_session,
+            display_name=f"{PACKAGE_NAME}-{session.name}",
+        )
+        .run_commands(opts.dev_run)
     )
-    session_run_commands(session, dev_run)
 
 
-# ** Dev (virtualenv)
-@DEFAULT_SESSION_VENV
-def dev_venv(
-    session: Session,
-    dev_run: RUN_CLI = [],  # noqa
-    lock: LOCK_CLI = False,
-    update: UPDATE_CLI = False,
-    update_package: UPDATE_PACKAGE_CLI = False,
-    log_session: bool = False,
-) -> None:
-    """Create dev env using virtualenv."""
-    # using conda
-
-    pkg_install_venv(
-        session=session,
-        name="dev-venv",
-        lock=lock,
-        requirement_paths="dev.txt",
-        display_name=f"{PACKAGE_NAME}-dev-venv",
-        install_package=True,
-        update=update,
-        update_package=update_package,
-        log_session=log_session,
-    )
-    session_run_commands(session, dev_run)
-
-
-# ** Notebooks
-@NOPYTHON_SESSION
-def notebook(
-    session: Session,
-    notebook_cmd: cmd_annotated(  # type: ignore
-        choices=["pytest", "mypy", "pyright", "all"], flags=("--notebook-cmd", "-n")
-    ) = [],  # noqa
-    notebooks: Annotated[
-        list[str], Option(nargs="*", type=str, help="notebooks to run on")
-    ] = [],  # noqa
-) -> None:
-    python = f".nox/{PACKAGE_NAME}/envs/dev/bin/python"
-
-    cmd = notebook_cmd
-    if not cmd or "all" in cmd:
-        cmd = ["mypy", "pyright", "pytest"]
-
-    if not notebooks:
-        notebooks = ["examples/usage"]
-
-    for c in cmd:
-        if c == "mypy":
-            session.run("nbqa", "--nbqa-shell", "mypy", *notebooks, external=True)
-        elif c == "pyright":
-            session.run(
-                "nbqa",
-                "--nbqa-shell",
-                "pyright",
-                *notebooks,
-                "--pythonpath",
-                python,
-                external=True,
-            )
-        elif c == "pytest":
-            session.run(
-                "pytest",
-                "--nbval",
-                "--current-env",
-                "--sanitize-with=config/nbval.ini",
-                *notebooks,
-                external=True,
-            )
+nox.session(name="dev-venv", **DEFAULT_KWS)(dev)
+nox.session(name="dev", **CONDA_DEFAULT_KWS)(dev)
 
 
 # ** bootstrap
-@NOPYTHON_SESSION
-def bootstrap(session: Session) -> None:
+@nox.session
+@add_opts
+def bootstrap(session: Session, opts: SessionParams) -> None:
     """Run config, reqs, and dev"""
+    session.install("nox", "ruamel.yaml")
 
-    session.notify("config")
-    session.notify("requirements")
-    session.notify("dev")
+    cmds = opts.bootstrap or ["config", "requirements", "dev"]
+
+    for c in cmds:
+        session.run("nox", "-s", c, "--", *session.posargs)
 
 
 # ** config
-@NOPYTHON_SESSION
+@nox.session(python=False)
+@add_opts
 def config(
     session: Session,
-    dev_extras: DEV_EXTRAS_CLI = [],  # type: ignore # noqa
-    python_paths: PYTHON_PATHS_CLI = [],  # type: ignore # noqa
+    opts: SessionParams,
 ) -> None:
     """Create the file ./config/userconfig.toml"""
-
-    args = []
-    if dev_extras:
-        args += ["--dev-extras"] + dev_extras
-    if python_paths:
-        args += ["--python-paths"] + python_paths
+    args: list[str] = []
+    if opts.dev_extras:
+        args += ["--dev-extras"] + opts.dev_extras
+    if opts.python_paths:
+        args += ["--python-paths"] + opts.python_paths
 
     session.run("python", "tools/projectconfig.py", *args)
 
 
 # ** requirements
-@NOPYTHON_SESSION
+@nox.session(python=False)
 def pyproject2conda(
     session: Session,
-    update: UPDATE_CLI = False,
 ) -> None:
     """Alias to reqs"""
+
     session.notify("requirements")
 
 
-@INHERITED_SESSION_VENV
+@nox.session
+@add_opts
 def requirements(
     session: Session,
-    update: UPDATE_CLI = False,
-    requirements_force: bool = False,
-    log_session: bool = False,
+    opts: SessionParams,
 ) -> None:
     """
     Create environment.yaml and requirement.txt files from pyproject.toml using pyproject2conda.
 
-    These will be placed in the directory "./environments".
+    These will be placed in the directory "./requirements".
     """
-    pkg_install_venv(
-        session=session,
-        reqs=["pyproject2conda>=0.8.0"],
-        name="reqs",
-        update=update,
-        log_session=log_session,
+
+    (
+        Installer(
+            session=session,
+            pip_deps="pyproject2conda>=0.11.0",
+            update=opts.update,
+        ).install_all(log_session=opts.log_session)
     )
 
     session.run(
         "pyproject2conda",
         "project",
         "--verbose",
-        *(["--overwrite", "force"] if requirements_force else []),
+        *(["--overwrite", "force"] if opts.requirements_force else []),
     )
 
+    if opts.lock:
+        for py in PYTHON_ALL_VERSIONS:
+            session.notify(f"pip-compile-{py}")
 
-# ** conda-lock
-@DEFAULT_SESSION_VENV
+
+# # ** conda-lock
+@nox.session(name="conda-lock", **DEFAULT_KWS)
+@add_opts
 def conda_lock(
     session: Session,
-    update: UPDATE_CLI = False,
-    conda_lock_channel: cmd_annotated(help="conda channels to use") = (),  # type: ignore
-    conda_lock_platform: cmd_annotated(  # type: ignore
-        help="platforms to build lock files for",
-        choices=["osx-64", "linux-64", "win-64", "all"],
-    ) = (),
-    conda_lock_include: cmd_annotated(  # type: ignore
-        help="lock files to create",
-    ) = (),
-    conda_lock_run: RUN_CLI = [],  # noqa
-    conda_lock_mamba: bool = False,
-    conda_lock_force: bool = False,
+    opts: SessionParams,
 ) -> None:
     """Create lock files using conda-lock."""
 
-    pkg_install_venv(
-        session,
-        name="conda-lock",
-        reqs=["conda-lock>=2.2.0"],
-        update=update,
+    (
+        Installer(
+            session=session,
+            pip_deps="conda-lock>=2.2.0",
+            update=opts.update,
+        ).install_all()
     )
 
     session.run("conda-lock", "--version")
 
-    conda_lock_exclude = ["test-extras"]
-    platform = cast(Sequence[str], conda_lock_platform)
-    if not platform:
-        platform = ["osx-64"]
-    elif "all" in platform:
-        platform = ["linux-64", "osx-64", "win-64"]
-    channel = cast(Sequence[str], conda_lock_channel)
+    conda_lock_exclude = ["test-extras", "dev"]
+
+    platform = opts.conda_lock_platform
+    if platform is None or "all" in platform:
+        # for now, skip osx-arm64 and win-64.  Leads to some problems.
+        platform = ["osx-64", "linux-64"]
+
+    channel = opts.conda_lock_channel
     if not channel:
         channel = ["conda-forge"]
 
@@ -376,139 +395,203 @@ def conda_lock(
 
         # check if skip
         env = "-".join(name.split("-")[1:])
-        if conda_lock_include:
-            if not any(c == env for c in conda_lock_include):
+        if opts.conda_lock_include:
+            if not any(c == env for c in opts.conda_lock_include):  # pyright: ignore
                 session.log(f"Skipping {lockfile} (include)")
                 return
 
         if conda_lock_exclude:
-            if any(c == env for c in conda_lock_exclude):
+            if any(c == env for c in conda_lock_exclude):  # pyright: ignore
                 session.log(f"Skipping {lockfile} (exclude)")
                 return
 
-        if conda_lock_force or update_target(lockfile, *deps):
+        # check hashes
+        changed, hash_path, hashes = check_hash_path_for_change(lockfile, *deps)
+        if opts.conda_lock_force or changed:
             session.log(f"Creating {lockfile}")
             # insert -f for each arg
             if lockfile.exists():
                 lockfile.unlink()
             session.run(
                 "conda-lock",
-                "--mamba" if conda_lock_mamba else "--no-mamba",
+                "--mamba" if opts.conda_lock_mamba else "--no-mamba",
                 *prepend_flag("-c", *channel),
                 *prepend_flag("-p", *platform),
                 *prepend_flag("-f", *deps),
                 f"--lockfile={lockfile}",
             )
+            write_hashes(hash_path, hashes)
         else:
             session.log(f"Skipping {lockfile} (exists)")
 
-    session_run_commands(session, conda_lock_run)
+    session_run_commands(session, opts.conda_lock_run)
     for path in (ROOT / "requirements").relative_to(ROOT.cwd()).glob("py*.yaml"):
         create_lock(path)
+
+
+@nox.session(name="pip-compile", **ALL_KWS)
+@add_opts
+def pip_compile(
+    session: Session,
+    opts: SessionParams,
+) -> None:
+    """Run pip-compile."""
+
+    (
+        Installer(
+            session=session, pip_deps=["pip-tools"], update=opts.update
+        ).install_all(log_session=opts.log_session)
+    )
+
+    envs_all = ["test", "typing"]
+    envs_dev = ["dev", "dev-complete", "dev-base", "docs", "test-notebook"]
+
+    if session.python == PYTHON_DEFAULT_VERSION:
+        envs = envs_all + envs_dev
+    else:
+        envs = envs_all
+
+    for env in envs:
+        assert isinstance(session.python, str)
+        reqspath = infer_requirement_path(env, ext=".txt")
+        lockpath = infer_requirement_path(
+            env,
+            ext=".txt",
+            python_version=session.python,
+            lock=True,
+            check_exists=False,
+        )
+
+        changed, hash_path, hashes = check_hash_path_for_change(lockpath, reqspath)
+        if opts.pip_compile_force or changed:
+            session.log(f"Creating {lockpath}")
+            session.run("pip-compile", "-U", "-o", lockpath, reqspath)
+            write_hashes(hash_path, hashes)
+
+        else:
+            session.log(f"Skipping {lockpath}")
 
 
 # ** testing
 def _test(
     session: nox.Session,
-    run: list[list[str]],
+    run: RUN_TYPE,
     test_no_pytest: bool,
-    test_opts: list[str],
+    test_opts: OPT_TYPE,
     no_cov: bool,
 ) -> None:
+    import os
+
+    tmpdir = os.environ.get("TMPDIR", None)
+
     session_run_commands(session, run)
     if not test_no_pytest:
-        opts = combine_list_str(test_opts)
+        opts = combine_list_str(test_opts or [])
         if not no_cov:
             session.env["COVERAGE_FILE"] = str(Path(session.create_tmp()) / ".coverage")
-            if "--cov" not in opts:
-                opts.append("--cov")
+
+            if not any(o.startswith("--cov") for o in opts):
+                opts.append(f"--cov={IMPORT_NAME}")
+
+        # Because we are testing if temporary folders
+        # have git or not, we have to make sure we're above the
+        # not under this repo
+        # so revert to using the top level `TMPDIR`
+        if tmpdir:
+            session.env["TMPDIR"] = tmpdir
+
         session.run("pytest", *opts)
 
 
-@ALL_SESSION
+# *** Basic tests
+@add_opts
 def test(
     session: Session,
-    test_no_pytest: bool = False,
-    test_opts: TEST_OPTS_CLI = (),  # type: ignore
-    test_run: RUN_CLI = [],  # noqa
-    lock: LOCK_CLI = False,
-    update: UPDATE_CLI = False,
-    update_package: UPDATE_PACKAGE_CLI = False,
-    log_session: bool = False,
-    no_cov: bool = False,
+    opts: SessionParams,
 ) -> None:
     """Test environments with conda installs."""
-
-    pkg_install_condaenv(
-        session=session,
-        name="test",
-        lock=lock,
-        install_package=True,
-        update=update,
-        update_package=update_package,
-        log_session=log_session,
+    (
+        Installer.from_envname(
+            session=session,
+            envname="test",
+            lock=opts.lock,
+            package=".",
+            update=opts.update,
+        ).install_all(log_session=opts.log_session, update_package=opts.update_package)
     )
 
     _test(
         session=session,
-        run=test_run,
-        test_no_pytest=test_no_pytest,
-        test_opts=test_opts,
-        no_cov=no_cov,
+        run=opts.test_run,
+        test_no_pytest=opts.test_no_pytest,
+        test_opts=opts.test_opts,
+        no_cov=opts.no_cov,
     )
 
 
-@ALL_SESSION_VENV
-def test_venv(
+nox.session(name="test", **ALL_KWS)(test)
+nox.session(name="test-conda", **CONDA_ALL_KWS)(test)
+
+
+@nox.session(name="test-notebook", **DEFAULT_KWS)
+@add_opts
+def test_notebook(session: nox.Session, opts: SessionParams) -> None:
+    (
+        Installer.from_envname(
+            session=session,
+            envname="test-notebook",
+            lock=opts.lock,
+            package=".",
+            update=opts.update,
+        ).install_all(log_session=opts.log_session, update_package=opts.update_package)
+    )
+
+    test_nbval_opts = shlex.split(
+        """
+    --nbval
+    --nbval-current-env
+    --nbval-sanitize-with=config/nbval.ini
+    --dist loadscope
+     examples/usage
+   """
+    )
+
+    test_opts = (opts.test_opts or []) + test_nbval_opts
+
+    session.log(f"{test_opts = }")
+
+    _test(
+        session=session,
+        run=opts.test_run,
+        test_no_pytest=opts.test_no_pytest,
+        test_opts=test_opts,
+        no_cov=opts.no_cov,
+    )
+
+
+# *** coverage
+@nox.session
+@add_opts
+def coverage(
     session: Session,
-    test_no_pytest: bool = False,
-    test_opts: TEST_OPTS_CLI = (),  # type: ignore
-    test_run: RUN_CLI = [],  # noqa
-    lock: LOCK_CLI = False,  # pyright: ignore
-    update: UPDATE_CLI = False,
-    update_package: UPDATE_PACKAGE_CLI = False,
-    log_session: bool = False,
-    no_cov: bool = False,
+    opts: SessionParams,
 ) -> None:
-    """Test environments virtualenv and pip installs."""
+    runner = Installer(
+        session=session, pip_deps="coverage[toml]", update=opts.update
+    ).install_all()
 
-    pkg_install_venv(
-        session=session,
-        name="test-venv",
-        install_package=True,
-        requirement_paths="test.txt",
-        update=update,
-        update_package=update_package,
-        log_session=log_session,
-    )
+    runner.run_commands(opts.coverage_run)
 
-    _test(
-        session=session,
-        run=test_run,
-        test_no_pytest=test_no_pytest,
-        test_opts=test_opts,
-        no_cov=no_cov,
-    )
-
-
-# ** coverage
-def _coverage(
-    session: nox.Session,
-    run: list[list[str]],
-    cmd: list[str],
-    run_internal: list[list[str]],
-) -> None:
-    session_run_commands(session, run)
-
-    if not cmd and not run and not run_internal:
-        cmd = ["combine", "html"]
+    cmd = opts.coverage or []
+    if not cmd and not opts.coverage_run and not opts.coverage_run_internal:
+        cmd = ["combine", "html", "report"]
 
     session.log(f"{cmd}")
 
     for c in cmd:
         if c == "combine":
             paths = list(
-                Path(session.virtualenv.location).parent.glob("test-3*/tmp/.coverage")
+                Path(session.virtualenv.location).parent.glob("test-*/tmp/.coverage")
             )
             if update_target(".coverage", *paths):
                 session.run("coverage", "combine", "--keep", "-a", *map(str, paths))
@@ -517,46 +600,76 @@ def _coverage(
         else:
             session.run("coverage", c)
 
-    session_run_commands(session, run_internal, external=False)
+    runner.run_commands(opts.coverage_run_internal)
 
 
-@INHERITED_SESSION_VENV
-def coverage(
+# *** testdist (conda)
+def testdist(
     session: Session,
-    coverage_cmd: cmd_annotated(  # type: ignore
-        choices=["erase", "combine", "report", "html", "open"]
-    ) = (),
-    coverage_run: RUN_CLI = [],  # noqa
-    coverage_run_internal: run_annotated(  # type: ignore
-        help="Arbitrary commands to run within the session"
-    ) = [],  # noqa
-    update: UPDATE_CLI = False,
 ) -> None:
-    pkg_install_venv(
-        session,
-        name="coverage",
-        reqs=["coverage[toml]"],
-        update=update,
+    """Test conda distribution."""
+
+    opts = parse_posargs(*session.posargs)
+
+    install_str = PACKAGE_NAME
+    if opts.version:
+        install_str = f"{install_str}=={opts.version}"
+
+    if is_conda_session(session):
+        pip_deps, conda_deps = None, install_str
+    else:
+        pip_deps, conda_deps = install_str, None
+
+    (
+        Installer.from_envname(
+            session=session,
+            envname="test-extras",
+            conda_deps=conda_deps,
+            pip_deps=pip_deps,
+            update=opts.update,
+            channels="conda-forge",
+        ).install_all(log_session=opts.log_session)
     )
 
-    _coverage(
+    _test(
         session=session,
-        run=coverage_run,
-        cmd=cast(list[str], coverage_cmd),
-        run_internal=cast(list[list[str]], coverage_run_internal),
+        run=opts.testdist_run,
+        test_no_pytest=opts.test_no_pytest,
+        test_opts=opts.test_opts,
+        no_cov=True,
     )
 
 
-# ** Docs
-def _docs(
-    session: nox.Session, run: list[list[str]], cmd: list[str], version: str
+nox.session(name="testdist-pypi", **ALL_KWS)(testdist)
+nox.session(name="testdist-conda", **CONDA_ALL_KWS)(testdist)
+
+
+# # ** Docs
+@add_opts
+def docs(
+    session: nox.Session,
+    opts: SessionParams,
 ) -> None:
-    if version:
-        session.env["SETUPTOOLS_SCM_PRETEND_VERSION"] = version
+    """
+    Runs make in docs directory. For example, 'nox -s docs -- +d html'
+    calls 'make -C docs html'. With 'release' option, you can set the
+    message with 'message=...' in posargs.
+    """
+    runner = Installer.from_envname(
+        session=session, envname="docs", lock=opts.lock, package=".", update=opts.update
+    ).install_all(
+        update_package=opts.update_package,
+        log_session=opts.log_session,
+        display_name=f"{PACKAGE_NAME}-{session.name}",
+    )
 
-    session_run_commands(session, run)
+    if opts.version:
+        session.env["SETUPTOOLS_SCM_PRETEND_VERSION"] = opts.version
 
-    if not run and not cmd:
+    runner.run_commands(opts.docs_run)
+
+    cmd = opts.docs or []
+    if not opts.docs_run and not cmd:
         cmd = ["html"]
 
     if "symlink" in cmd:
@@ -590,300 +703,16 @@ def _docs(
         )
 
 
-@DEFAULT_SESSION
-def docs(
-    session: nox.Session,
-    docs_cmd: cmd_annotated(  # type: ignore
-        choices=[
-            "html",
-            "build",
-            "symlink",
-            "clean",
-            "livehtml",
-            "linkcheck",
-            "spelling",
-            "showlinks",
-            "release",
-            "open",
-            "serve",
-        ],
-        flags=("--docs-cmd", "-d"),
-    ) = (),
-    docs_run: RUN_CLI = [],  # noqa
-    lock: LOCK_CLI = False,
-    update: UPDATE_CLI = False,
-    update_package: UPDATE_PACKAGE_CLI = False,
-    version: VERSION_CLI = "",
-    log_session: bool = False,
-) -> None:
-    """Runs make in docs directory. For example, 'nox -s docs -- --docs-cmd html' -> 'make -C docs html'. With 'release' option, you can set the message with 'message=...' in posargs."""
-    pkg_install_condaenv(
-        session=session,
-        name="docs",
-        lock=lock,
-        display_name=f"{PACKAGE_NAME}-docs",
-        install_package=True,
-        update=update,
-        update_package=update_package,
-        log_session=log_session,
-    )
-
-    _docs(
-        session=session, cmd=docs_cmd, run=docs_run, version=version
-    )  # pyright: ignore
-
-
-@DEFAULT_SESSION_VENV
-def docs_venv(
-    session: nox.Session,
-    docs_cmd: cmd_annotated(  # type: ignore
-        choices=[
-            "html",
-            "build",
-            "symlink",
-            "clean",
-            "livehtml",
-            "linkcheck",
-            "spelling",
-            "showlinks",
-            "release",
-            "open",
-            "serve",
-        ],
-        flags=("--docs-cmd", "-d"),
-    ) = (),
-    docs_run: RUN_CLI = [],  # noqa
-    lock: LOCK_CLI = False,
-    update: UPDATE_CLI = False,
-    update_package: UPDATE_PACKAGE_CLI = False,
-    version: VERSION_CLI = "",
-    log_session: bool = False,
-) -> None:
-    """Runs make in docs directory. For example, 'nox -s docs -- --docs-cmd html' -> 'make -C docs html'. With 'release' option, you can set the message with 'message=...' in posargs."""
-    pkg_install_venv(
-        session=session,
-        name="docs-venv",
-        lock=lock,
-        display_name=f"{PACKAGE_NAME}-docs-venv",
-        install_package=True,
-        update=update,
-        update_package=update_package,
-        log_session=log_session,
-        requirement_paths="docs.txt",
-    )
-
-    _docs(
-        session=session, cmd=docs_cmd, run=docs_run, version=version
-    )  # pyright: ignore
-
-
-# ** Dist pypi
-def _dist_pypi(
-    session: nox.Session, run: list[list[str]], cmd: list[str], version: str
-) -> None:
-    if version:
-        session.env["SETUPTOOLS_SCM_PRETEND_VERSION"] = version
-    session_run_commands(session, run)
-    if not run and not cmd:
-        cmd = ["build"]
-    if cmd:
-        if "build" in cmd:
-            cmd.append("clean")
-        cmd = sort_like(cmd, ["clean", "build", "testrelease", "release"])
-
-        session.log(f"cmd={cmd}")
-
-        for command in cmd:
-            if command == "clean":
-                session.run("rm", "-rf", "dist", external=True)
-            elif command == "build":
-                session.run("python", "-m", "build", "--outdir", "dist/")
-
-            elif command == "testrelease":
-                session.run("twine", "upload", "--repository", "testpypi", "dist/*")
-
-            elif command == "release":
-                session.run("twine", "upload", "dist/*")
-
-
-@DEFAULT_SESSION_VENV
-def dist_pypi(
-    session: nox.Session,
-    dist_pypi_run: RUN_CLI = [],  # noqa
-    dist_pypi_cmd: cmd_annotated(  # type: ignore
-        choices=["clean", "build", "testrelease", "release"],
-        flags=("--dist-pypi-cmd", "-p"),
-    ) = (),
-    lock: LOCK_CLI = False,  # pyright: ignore
-    update: UPDATE_CLI = False,
-    version: VERSION_CLI = "",
-    log_session: bool = False,
-) -> None:
-    """Run 'nox -s dist-pypi -- {clean, build, testrelease, release}'."""
-
-    pkg_install_venv(
-        session=session,
-        name="dist-pypi",
-        requirement_paths="dist-pypi.txt",
-        update=update,
-        install_package=False,
-        log_session=log_session,
-    )
-
-    _dist_pypi(
-        session=session,
-        run=dist_pypi_run,
-        cmd=dist_pypi_cmd,  # pyright: ignore
-        version=version,
-    )
-
-
-@DEFAULT_SESSION
-def dist_pypi_condaenv(
-    session: nox.Session,
-    dist_pypi_run: RUN_CLI = [],  # noqa
-    dist_pypi_cmd: cmd_annotated(  # type: ignore
-        choices=["clean", "build", "testrelease", "release"],
-        flags=("--dist-pypi-cmd", "-p"),
-    ) = (),
-    lock: LOCK_CLI = False,  # pyright: ignore
-    update: UPDATE_CLI = False,
-    version: VERSION_CLI = "",
-    log_session: bool = False,
-) -> None:
-    """Run 'nox -s dist_pypi -- {clean, build, testrelease, release}'."""
-    # conda
-
-    pkg_install_condaenv(
-        session=session,
-        name="dist-pypi",
-        install_package=False,
-        update=update,
-        log_session=log_session,
-    )
-
-    _dist_pypi(
-        session=session,
-        run=dist_pypi_run,
-        cmd=dist_pypi_cmd,  # pyright: ignore
-        version=version,
-    )
-
-
-# ** Dist conda
-@DEFAULT_SESSION
-def dist_conda(
-    session: nox.Session,
-    dist_conda_run: RUN_CLI = [],  # noqa
-    dist_conda_cmd: cmd_annotated(  # type: ignore
-        choices=[
-            "recipe",
-            "build",
-            "clean",
-            "clean-recipe",
-            "clean-build",
-            "recipe-cat-full",
-        ],
-        flags=("--dist-conda-cmd", "-c"),
-    ) = (),
-    # lock: LOCK_CLI = False,
-    sdist_path: str = "",
-    update: UPDATE_CLI = False,
-    log_session: bool = False,
-    version: VERSION_CLI = "",
-) -> None:
-    """Runs make -C dist-conda posargs."""
-    pkg_install_condaenv(
-        session=session,
-        name="dist-conda",
-        install_package=False,
-        update=update,
-        log_session=log_session,
-    )
-
-    run, cmd = dist_conda_run, dist_conda_cmd
-    session_run_commands(session, run)
-    if not run and not cmd:
-        cmd = ["recipe"]
-
-    # make directory?
-    if not (d := Path("./dist-conda")).exists():
-        d.mkdir()
-
-    if cmd:
-        if "recipe" in cmd:
-            cmd.append("clean-recipe")
-        if "build" in cmd:
-            cmd.append("clean-build")
-        if "clean" in cmd:
-            cmd.extend(["clean-recipe", "clean-build"])
-            cmd.remove("clean")
-
-        cmd = sort_like(
-            cmd, ["recipe-cat-full", "clean-recipe", "recipe", "clean-build", "build"]
-        )
-
-        if not sdist_path:
-            sdist_path = PACKAGE_NAME
-            if version:
-                sdist_path = f"{sdist_path}=={version}"
-
-        for command in cmd:
-            if command == "clean-recipe":
-                session.run("rm", "-rf", f"dist-conda/{PACKAGE_NAME}", external=True)
-            elif command == "clean-build":
-                session.run("rm", "-rf", "dist-conda/build", external=True)
-            elif command == "recipe":
-                session.run(
-                    "grayskull",
-                    "pypi",
-                    sdist_path,
-                    "--sections",
-                    "package",
-                    "source",
-                    "build",
-                    "requirements",
-                    "-o",
-                    "dist-conda",
-                )
-                _append_recipe(
-                    f"dist-conda/{PACKAGE_NAME}/meta.yaml", "config/recipe-append.yaml"
-                )
-                session.run(
-                    "cat", f"dist-conda/{PACKAGE_NAME}/meta.yaml", external=True
-                )
-            elif command == "recipe-cat-full":
-                import tempfile
-
-                with tempfile.TemporaryDirectory() as d:  # type: ignore[assignment,unused-ignore]
-                    session.run(
-                        "grayskull",
-                        "pypi",
-                        sdist_path,
-                        "-o",
-                        str(d),
-                    )
-                    session.run(
-                        "cat", str(Path(d) / PACKAGE_NAME / "meta.yaml"), external=True
-                    )
-
-            elif command == "build":
-                session.run(
-                    "conda",
-                    "mambabuild",
-                    "--output-folder=dist-conda/build",
-                    "--no-anaconda-upload",
-                    "dist-conda",
-                )
+nox.session(name="docs", **DEFAULT_KWS)(docs)
+nox.session(name="docs-conda", **CONDA_DEFAULT_KWS)(docs)
 
 
 # ** lint
-@INHERITED_SESSION_VENV
+@nox.session
+@add_opts
 def lint(
     session: nox.Session,
-    lint_run: RUN_CLI = [],  # noqa
-    update: UPDATE_CLI = False,
-    log_session: bool = False,
+    opts: SessionParams,
 ) -> None:
     """
     Run linters with pre-commit.
@@ -892,30 +721,38 @@ def lint(
     To run something else pass, e.g.,
     `nox -s lint -- --lint-run "pre-commit run --hook-stage manual --all-files`
     """
-    pkg_install_venv(
-        session=session,
-        name="lint",
-        reqs=["pre-commit"],
-        install_package=False,
-        update=update,
-        log_session=log_session,
-    )
+    runner = Installer(
+        session=session, pip_deps="pre-commit", update=opts.update
+    ).install_all(log_session=opts.log_session)
 
-    if lint_run:
-        session_run_commands(session, lint_run, external=False)
+    if opts.lint_run:
+        runner.run_commands(opts.lint_run, external=True)
     else:
         session.run("pre-commit", "run", "--all-files")
 
 
 # ** type checking
-def _typing(
+@add_opts
+def typing(
     session: nox.Session,
-    run: list[list[str]],
-    cmd: list[str],
-    run_internal: list[list[str]],
+    opts: SessionParams,
 ) -> None:
-    session_run_commands(session, run)
-    if not run and not run_internal and not cmd:
+    """Run type checkers (mypy, pyright, pytype)."""
+    runner = (
+        Installer.from_envname(
+            session=session,
+            envname="typing",
+            lock=opts.lock,
+            update=opts.update,
+            # need package for nbqa checks
+            package=".",
+        )
+        .install_all(log_session=opts.log_session)
+        .run_commands(opts.typing_run)
+    )
+
+    cmd = opts.typing or []
+    if not opts.typing_run and not opts.typing_run_internal and not cmd:
         cmd = ["mypy", "pyright"]
 
     if "all" in cmd:
@@ -929,7 +766,7 @@ def _typing(
         session.run(cmd, "--version", external=True)
 
     if "clean" in cmd:
-        cmd = [x for x in cmd if x != "clean"]
+        cmd.remove("clean")
 
         for name in [".mypy_cache", ".pytype"]:
             p = Path(session.create_tmp()) / name
@@ -938,226 +775,200 @@ def _typing(
                 shutil.rmtree(str(p))
 
     for c in cmd:
-        if not c.startswith("nbqa"):
+        if c.endswith("-notebook"):
+            session.run("make", c, external=True)
+            continue
+        else:
             _run_info(c)
+
         if c == "mypy":
             session.run("mypy", "--color-output")
         elif c == "pyright":
             session.run("pyright", external=True)
         elif c == "pytype":
             session.run("pytype", "-o", str(Path(session.create_tmp()) / ".pytype"))
-        elif c.startswith("nbqa"):
-            session.run("make", c, external=True)
         else:
             session.log(f"skipping unknown command {c}")
-    session_run_commands(session, run_internal, external=False)
+
+    runner.run_commands(opts.typing_run_internal, external=False)
 
 
-@ALL_SESSION
-def typing(
+nox.session(name="typing", **ALL_KWS)(typing)
+nox.session(name="typing-conda", **CONDA_ALL_KWS)(typing)
+
+
+# # ** Dist pypi
+@nox.session
+@add_opts
+def build(session: nox.Session, opts: SessionParams) -> None:
+    """
+    Build the distribution.
+
+    Note that default is to not use build isolation.
+    Pass `--build-isolation` to use build isolation.
+    """
+    (
+        Installer.from_envname(
+            session=session,
+            envname="build",
+            update=opts.update,
+        ).install_all(log_session=opts.log_session)
+    )
+
+    if opts.version:
+        session.env["SETUPTOOLS_SCM_PRETEND_VERSION"] = opts.version
+
+    for cmd in opts.build or ["build"]:
+        if cmd == "version":
+            session.run("python", "-m", "setuptools_scm")
+
+        elif cmd == "build":
+            if Path("dist").exists():
+                shutil.rmtree("./dist")
+
+            args = "python -m build --outdir dist".split()
+            if not opts.build_isolation:
+                args.append("--no-isolation")
+
+            session.run(*args)
+
+
+@nox.session
+@add_opts
+def publish(session: nox.Session, opts: SessionParams) -> None:
+    """Publish the distribution"""
+
+    (
+        Installer(session=session, pip_deps="twine", update=opts.update)
+        .install_all(log_session=opts.log_session)
+        .run_commands(opts.publish_run)
+    )
+
+    for cmd in opts.publish or []:
+        if cmd == "test":
+            session.run("twine", "upload", "--repository", "testpypi", "dist/*")
+
+        elif cmd == "release":
+            session.run("twine", "upload", "dist/*")
+
+
+# # ** Dist conda
+@nox.session(name="conda-recipe")
+@add_opts
+def conda_recipe(
     session: nox.Session,
-    typing_cmd: cmd_annotated(  # type: ignore
-        choices=[
-            "clean",
-            "mypy",
-            "pyright",
-            "pytype",
-            "all",
-            "nbqa-mypy",
-            "nbqa-pyright",
-            "nbqa-typing",
-        ],
-        flags=("--typing-cmd", "-m"),
-    ) = (),
-    typing_run: RUN_CLI = [],  # noqa
-    typing_run_internal: run_annotated(  # type: ignore
-        help="run arbitrary (internal) commands.  For example, --typing-run-internal 'mypy --some-option'",
-    ) = [],  # noqa
-    lock: LOCK_CLI = False,
-    update: UPDATE_CLI = False,
-    log_session: bool = False,
+    opts: SessionParams,
 ) -> None:
-    """Run type checkers (mypy, pyright, pytype)."""
+    """Run grayskull to create recipe"""
+    runner = Installer(
+        session=session, pip_deps=["grayskull"], update=opts.update
+    ).install_all(log_session=opts.log_session)
 
-    pkg_install_condaenv(
-        session=session,
-        name="typing",
-        lock=lock,
-        install_package=False,
-        update=update,
-        log_session=log_session,
-    )
+    run, commands = opts.conda_recipe_run, opts.conda_recipe
 
-    _typing(
-        session=session,
-        run=typing_run,
-        cmd=typing_cmd,
-        run_internal=typing_run_internal,
-    )
+    runner.run_commands(run)
 
+    if not run and not commands:
+        commands = ["recipe"]
 
-@ALL_SESSION_VENV
-def typing_venv(
-    session: nox.Session,
-    typing_cmd: cmd_annotated(  # type: ignore
-        choices=[
-            "clean",
-            "mypy",
-            "pyright",
-            "pytype",
-            "all",
-            "nbqa-mypy",
-            "nbqa-pyright",
-            "nbqa-typing",
-        ],
-        flags=("--typing-cmd", "-m"),
-    ) = (),
-    typing_run: RUN_CLI = [],  # noqa
-    typing_run_internal: run_annotated(  # type: ignore
-        help="run arbitrary (internal) commands.  For example, --typing-run-internal 'mypy --some-option'",
-    ) = [],  # noqa
-    lock: LOCK_CLI = False,
-    update: UPDATE_CLI = False,
-    log_session: bool = False,
-) -> None:
-    """Run type checkers (mypy, pyright, pytype)."""
+    if not commands:
+        return
 
-    pkg_install_venv(
-        session=session,
-        name="typing",
-        lock=lock,
-        requirement_paths="typing.txt",
-        install_package=False,
-        update=update,
-        log_session=log_session,
-    )
+    sdist_path = opts.conda_recipe_sdist_path
+    if not sdist_path:
+        sdist_path = PACKAGE_NAME
+        if opts.version:
+            sdist_path = f"{sdist_path}=={opts.version}"
 
-    _typing(
-        session=session,
-        run=typing_run,
-        cmd=typing_cmd,
-        run_internal=typing_run_internal,
-    )
+    for command in commands:
+        if command == "recipe":
+            # make directory?
+            if not (d := Path("./dist-conda")).exists():
+                d.mkdir()
+
+            session.run(
+                "grayskull",
+                "pypi",
+                sdist_path,
+                "--sections",
+                "package",
+                "source",
+                "build",
+                "requirements",
+                "-o",
+                "dist-conda",
+            )
+            _append_recipe(
+                f"dist-conda/{PACKAGE_NAME}/meta.yaml", "config/recipe-append.yaml"
+            )
+            session.run("cat", f"dist-conda/{PACKAGE_NAME}/meta.yaml", external=True)
+
+        elif command == "recipe-full":
+            import tempfile
+
+            with tempfile.TemporaryDirectory() as d:  # type: ignore[assignment,unused-ignore]
+                session.run(
+                    "grayskull",
+                    "pypi",
+                    sdist_path,
+                    "-o",
+                    str(d),
+                )
+                session.run(
+                    "cat", str(Path(d) / PACKAGE_NAME / "meta.yaml"), external=True
+                )
 
 
-# ** testdist (conda)
-@ALL_SESSION
-def testdist_conda(
-    session: Session,
-    test_no_pytest: bool = False,
-    test_opts: TEST_OPTS_CLI = (),  # type: ignore
-    testdist_conda_run: RUN_CLI = [],  # noqa
-    update: UPDATE_CLI = False,
-    version: VERSION_CLI = "",
-    log_session: bool = False,
-) -> None:
-    """Test conda distribution."""
+@nox.session(name="conda-build", **CONDA_DEFAULT_KWS)
+@add_opts
+def conda_build(session: nox.Session, opts: SessionParams) -> None:
+    runner = Installer.from_envname(
+        session=session, update=opts.update, conda_deps=["boa", "anaconda-client"]
+    ).install_all(log_session=opts.log_session)
 
-    install_str = PACKAGE_NAME
-    if version:
-        install_str = f"{install_str}=={version}"
+    cmds, run = opts.conda_build, opts.conda_build_run
 
-    pkg_install_condaenv(
-        session=session,
-        name="test-extras",
-        deps=[install_str],
-        channels=["conda-forge"],
-        update=update,
-        install_package=False,
-        log_session=log_session,
-    )
+    runner.run_commands(run)
 
-    _test(
-        session=session,
-        run=testdist_conda_run,
-        test_no_pytest=test_no_pytest,
-        test_opts=test_opts,
-        no_cov=True,
-    )
+    if not run and not cmds:
+        cmds = ["build", "clean"]
 
+    if cmds is None:
+        cmds = []
 
-# ** testdist (pypi)
-@ALL_SESSION_VENV
-def testdist_pypi(
-    session: Session,
-    test_no_pytest: bool = False,
-    test_opts: TEST_OPTS_CLI = (),  # type: ignore
-    testdist_pypi_run: RUN_CLI = [],  # noqa
-    testdist_pypi_extras: cmd_annotated(help="extras to install") = (),  # type: ignore
-    update: UPDATE_CLI = False,
-    version: VERSION_CLI = "",
-    log_session: bool = False,
-) -> None:
-    """Test pypi distribution."""
-    extras = testdist_pypi_extras
-    install_str = PACKAGE_NAME
+    if "clean" in cmds:
+        cmds.remove("clean")
+        session.log("removing directory dist-conda/build")
+        if Path("./dist-conda/build").exists():
+            shutil.rmtree("./dist-conda/build")
 
-    if extras:
-        install_str = "{}[{}]".format(install_str, ",".join(extras))
+    for cmd in cmds:
+        if cmd == "build":
+            if not (d := Path(f"./dist-conda/{PACKAGE_NAME}/meta.yaml")).exists():
+                raise ValueError(f"no file {d}")
 
-    if version:
-        install_str = f"{install_str}=={version}"
-
-    pkg_install_venv(
-        session=session,
-        name="testdist-pypi",
-        requirement_paths="test-extras.txt",
-        reqs=[install_str],
-        update=update,
-        install_package=False,
-        log_session=log_session,
-    )
-
-    _test(
-        session=session,
-        run=testdist_pypi_run,
-        test_no_pytest=test_no_pytest,
-        test_opts=test_opts,
-        no_cov=True,
-    )
+            session.run(
+                "conda",
+                "mambabuild",
+                "--output-folder=dist-conda/build",
+                "--no-anaconda-upload",
+                "dist-conda",
+            )
 
 
-@ALL_SESSION
-def testdist_pypi_condaenv(
-    session: Session,
-    test_no_pytest: bool = False,
-    test_opts: TEST_OPTS_CLI = (),  # type: ignore
-    testdist_pypi_run: RUN_CLI = [],  # noqa
-    testdist_pypi_extras: cmd_annotated(help="extras to install") = (),  # type: ignore
-    update: UPDATE_CLI = False,
-    version: VERSION_CLI = "",
-    log_session: bool = False,
-) -> None:
-    """Test pypi distribution."""
-    extras = testdist_pypi_extras
-    install_str = PACKAGE_NAME
+# ** Other utilities
+@nox.session
+@add_opts
+def cog(session: nox.Session, opts: SessionParams) -> None:
+    """Run cog."""
 
-    if extras:
-        install_str = "{}[{}]".format(install_str, ",".join(extras))
-
-    if version:
-        install_str = f"{install_str}=={version}"
-
-    pkg_install_condaenv(
-        session=session,
-        name="test-extras",
-        reqs=[install_str],
-        channels=["conda-forge"],
-        update=update,
-        install_package=False,
-        log_session=log_session,
-    )
-
-    _test(
-        session=session,
-        run=testdist_pypi_run,
-        test_no_pytest=test_no_pytest,
-        test_opts=test_opts,
-        no_cov=True,
-    )
+    Installer.from_envname(
+        session=session, update=opts.update, pip_deps="cogapp"
+    ).install_all(log_session=opts.log_session)
+    session.run("cog", "-rP", "README.md", env={"COLUMNS": "90"})
 
 
-# * Utilities --------------------------------------------------------------------------
+# * Utilities -------------------------------------------------------------------------
 def _create_doc_examples_symlinks(session: nox.Session, clean: bool = True) -> None:
     """Create symlinks from docs/examples/*.md files to /examples/usage/..."""
 
@@ -1228,7 +1039,7 @@ def _append_recipe(recipe_path: str, append_path: str) -> None:
 # # If want separate env for updating/reporting version with setuptools-scm
 # # We do this from dev environment.
 # # ** version report/update
-# @DEFAULT_SESSION_VENV
+# @nox.session(**DEFAULT_KWS)
 # def version_scm(
 #     session: Session,
 #     version: VERSION_CLI = "",
@@ -1247,7 +1058,7 @@ def _append_recipe(recipe_path: str, append_path: str) -> None:
 #         name="version-scm",
 #         install_package=True,
 #         reqs=["setuptools_scm"],
-#         update=update,
+#         update=opts.update,
 #         no_deps=True,
 #     )
 
